requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docutils==0.21.2
idna==3.7
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
paho-mqtt==2.1.0
Pygments==2.18.0
pypiwin32==223
pywin32==306
requests==2.32.3
urllib3==2.2.1
